# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: VNSF Scholarship Management API
  version: 0.1.0
  license: { name: Proprietary, identifier: LicenseRef-Proprietary }
servers:
  - url: /api/v1
security: [{ sessionCookie: [] }]
paths:
  /auth/login:
    post:
      summary: Authenticate and start a rotated session
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password:
                  {
                    type: string,
                    minLength: 8,
                    maxLength: 200,
                    format: password,
                  }
      responses:
        "200": { description: Session cookie and MFA requirement }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { description: Actor or IP rate limit exceeded }
  /auth/mfa/verify:
    post:
      summary: Verify TOTP for a pending administrative session
      operationId: verifyMfa
      security: [{ sessionCookie: [], csrfToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [code]
              properties: { code: { type: string, pattern: "^\\d{6}$" } }
      responses:
        "200": { description: MFA verified }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /auth/logout:
    post:
      summary: Revoke the current session
      operationId: logout
      security: [{ sessionCookie: [], csrfToken: [] }]
      responses:
        "200": { description: Session revoked }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /auth/sessions:
    get:
      summary: List active sessions owned by the current user
      operationId: listSessions
      responses:
        "200": { description: Active sessions without token material }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /auth/sessions/{id}:
    delete:
      summary: Revoke an owned session
      operationId: revokeSession
      security: [{ sessionCookie: [], csrfToken: [] }]
      parameters: [{ $ref: "#/components/parameters/ResourceId" }]
      responses:
        "200": { description: Session revoked }
        "404": { $ref: "#/components/responses/NotFound" }
  /auth/password/forgot:
    post:
      summary: Request password reset without disclosing account existence
      operationId: requestPasswordReset
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [email]
              properties: { email: { type: string, format: email } }
      responses:
        "200": { description: Request accepted regardless of account existence }
        "429": { description: Actor or IP rate limit exceeded }
  /auth/password/reset:
    post:
      summary: Consume a one-time password reset token
      operationId: resetPassword
      security: []
      requestBody: { $ref: "#/components/requestBodies/TokenPassword" }
      responses:
        "200": { description: Password changed and sessions revoked }
        "422": { $ref: "#/components/responses/ValidationFailed" }
  /auth/activate:
    post:
      summary: Activate an account using a one-time token
      operationId: activateAccount
      security: []
      requestBody: { $ref: "#/components/requestBodies/TokenPassword" }
      responses:
        "200": { description: Account activated }
        "422": { $ref: "#/components/responses/ValidationFailed" }
  /auth/reauthenticate:
    post:
      summary: Re-authenticate before a sensitive operation
      operationId: reauthenticate
      security: [{ sessionCookie: [], csrfToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [password]
              properties:
                {
                  password:
                    {
                      type: string,
                      minLength: 12,
                      maxLength: 200,
                      format: password,
                    },
                }
      responses:
        "200": { description: Re-authentication valid for five minutes }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /auth/mfa/enroll:
    post:
      summary: Begin TOTP enrollment for the current user
      operationId: beginMfaEnrollment
      security: [{ sessionCookie: [], csrfToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [email]
              properties: { email: { type: string, format: email } }
      responses:
        "200": { description: TOTP provisioning URI }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /auth/mfa/enroll/confirm:
    post:
      summary: Confirm TOTP enrollment and issue recovery codes once
      operationId: confirmMfaEnrollment
      security: [{ sessionCookie: [], csrfToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [code]
              properties: { code: { type: string, pattern: "^\\d{6}$" } }
      responses:
        "200": { description: One-time recovery code list }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /auth/mfa/recovery/verify:
    post:
      summary: Consume one MFA recovery code
      operationId: verifyMfaRecoveryCode
      security: [{ sessionCookie: [], csrfToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [code]
              properties:
                { code: { type: string, minLength: 8, maxLength: 64 } }
      responses:
        "200": { description: MFA verified and recovery code consumed }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /configuration/{kind}:
    parameters:
      - in: path
        name: kind
        required: true
        schema: { type: string, enum: [schools, programs, periods, calendar] }
    get:
      summary: List operational configuration records
      operationId: listConfiguration
      responses:
        "200":
          {
            description: Configuration records visible to an authorized manager,
          }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      summary: Create an operational configuration record
      operationId: createConfiguration
      security: [{ sessionCookie: [], csrfToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ConfigurationInput" }
      responses:
        "200": { description: Created configuration record }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
  /configuration/{kind}/{id}:
    patch:
      summary: Optimistically update an operational configuration record
      operationId: updateConfiguration
      security: [{ sessionCookie: [], csrfToken: [] }]
      parameters:
        - in: path
          name: kind
          required: true
          schema: { type: string, enum: [schools, programs, periods, calendar] }
        - { $ref: "#/components/parameters/ResourceId" }
        - { $ref: "#/components/parameters/IfMatch" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ConfigurationInput" }
      responses:
        "200": { description: Updated configuration record }
        "404": { $ref: "#/components/responses/NotFound" }
        "412": { $ref: "#/components/responses/Conflict" }
        "428": { $ref: "#/components/responses/PreconditionRequired" }
...
  /submissions/{id}/submit:
    post:
      summary: Submit an immutable academic version
      operationId: submitAcademicSubmission
      security: [{ sessionCookie: [], csrfToken: [] }]
      parameters:
        [
          { $ref: "#/components/parameters/ResourceId" },
          { $ref: "#/components/parameters/IdempotencyKey" },
        ]
      requestBody:
        {
          required: true,
          content:
            {
              application/json:
                {
                  schema:
                    {
                      type: object,
                      required: [version],
                      properties: { version: { type: integer, minimum: 1 } },
                    },
                },
            },
        }
      responses:
        "200": { description: Immutable submitted version and receipt }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
  /submissions:
    get:
      summary: List submissions in the caller student or review scope
      operationId: listAcademicSubmissions
      parameters:
        [{ in: query, name: queue, schema: { type: boolean, default: false } }]
      responses:
        "200": { description: Scoped submissions or open review queue }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      summary: Create a unique draft submission
      operationId: createAcademicSubmission
      security: [{ sessionCookie: [], csrfToken: [] }]
      requestBody:
        required: true
        content:
          {
            application/json:
              {
                schema: { $ref: "#/components/schemas/SubmissionCreateInput" },
              },
          }
      responses:
        "200": { description: Draft submission created }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
  /submissions/{id}/draft:
    patch:
      summary: Optimistically save mutable draft data
      operationId: saveAcademicDraft
      security: [{ sessionCookie: [], csrfToken: [] }]
      parameters:
        - { $ref: "#/components/parameters/ResourceId" }
        - { $ref: "#/components/parameters/IfMatch" }
      requestBody:
        required: true
        content:
          {
            application/json:
              { schema: { $ref: "#/components/schemas/SubmissionDraftInput" } },
          }
      responses:
        "200": { description: Draft saved }
        "409": { $ref: "#/components/responses/Conflict" }
        "412": { $ref: "#/components/responses/Conflict" }
        "428": { $ref: "#/components/responses/PreconditionRequired" }
  /submissions/{id}/review:
    post:
      summary: Complete the current scoped review task
      operationId: reviewAcademicSubmission
      security: [{ sessionCookie: [], csrfToken: [] }]
      parameters:
        - { $ref: "#/components/parameters/ResourceId" }
        - { $ref: "#/components/parameters/IfMatch" }
      requestBody:
        required: true
        content:
          {
            application/json:
              {
                schema: { $ref: "#/components/schemas/SubmissionReviewInput" },
              },
          }
      responses:
        "200": { description: Review decision applied }
        "409": { $ref: "#/components/responses/Conflict" }
        "412": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
...
  /documents/upload-init:
    post:
      summary: Create a private quarantine upload with a short-lived signed URL
      operationId: initiateDocumentUpload
      security: [{ sessionCookie: [], csrfToken: [] }]
      requestBody:
        required: true
        content:
          {
            application/json:
              { schema: { $ref: "#/components/schemas/DocumentUploadInput" } },
          }
      responses:
        "200":
          { description: Signed quarantine upload URL and required headers }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
  /documents/{id}/complete:
    post:
      summary: Verify object metadata and magic bytes then enqueue antivirus scan
      operationId: completeDocumentUpload
      security: [{ sessionCookie: [], csrfToken: [] }]
      parameters: [{ $ref: "#/components/parameters/ResourceId" }]
      responses:
        "200": { description: Document pending antivirus scan }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
  /documents/{id}/download:
    get:
      summary: Get a short-lived download URL for an authorized clean document
      operationId: downloadDocument
      parameters: [{ $ref: "#/components/parameters/ResourceId" }]
      responses:
        "200": { description: Five-minute private download URL }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: vnsf_session
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
  requestBodies:
    TokenPassword:
      required: true
      content:
        application/json:
          schema:
            type: object
            additionalProperties: false
            required: [token, password]
            properties:
              token: { type: string, minLength: 32, maxLength: 200 }
              password:
                {
                  type: string,
                  minLength: 12,
                  maxLength: 200,
                  format: password,
                }
  parameters:
...
    ResourceId:
      {
        in: path,
        name: id,
        required: true,
        schema: { type: string, format: uuid },
      }
    IfMatch:
      {
        in: header,
        name: If-Match,
        required: true,
        schema: { type: string },
        example: '"3"',
      }
    IdempotencyKey:
      {
        in: header,
        name: Idempotency-Key,
        required: true,
        schema: { type: string, minLength: 16, maxLength: 128 },
      }
  schemas:
    DocumentUploadInput:
      type: object
      additionalProperties: false
      required:
        [
          owner_type,
          owner_id,
          purpose,
          filename,
          size_bytes,
          mime_type,
          checksum_sha256,
        ]
      properties:
        owner_type:
          { type: string, enum: [STUDENT, SUBMISSION, EDUCATION_EXPENSE] }
        owner_id: { type: string, format: uuid }
        purpose: { type: string, minLength: 2, maxLength: 40 }
        filename: { type: string, minLength: 1, maxLength: 255 }
        size_bytes: { type: integer, minimum: 1, maximum: 10485760 }
        mime_type:
          { type: string, enum: [application/pdf, image/jpeg, image/png] }
        checksum_sha256: { type: string, pattern: "^[a-f0-9]{64}$" }
    SubmissionCreateInput:
      type: object
      additionalProperties: false
      required: [student_id, period_id, type]
      properties:
        student_id: { type: string, format: uuid }
        period_id: { type: string, format: uuid }
        type: { type: string, minLength: 2, maxLength: 30 }
    SubmissionDraftInput:
      type: object
      additionalProperties: false
      required: [payload]
      properties: { payload: { type: object, additionalProperties: true } }
    SubmissionReviewInput:
      type: object
      additionalProperties: false
      required: [decision]
      properties:
        decision: { type: string, enum: [APPROVE, RETURN, REJECT] }
        reason_code: { type: string, minLength: 2, maxLength: 50 }
        note: { type: string, maxLength: 2000 }
    ConfigurationInput:
      oneOf:
        - type: object
          additionalProperties: false
          required: [code, name]
          properties:
            code: { type: string, minLength: 2, maxLength: 30 }
            name: { type: string, minLength: 2, maxLength: 150 }
            active: { type: boolean }
        - type: object
          additionalProperties: false
          required: [code, name, workflow_type]
          properties:
            code: { type: string, minLength: 2, maxLength: 30 }
            name: { type: string, minLength: 2, maxLength: 150 }
            workflow_type: { type: string, enum: [ONE_LEVEL, TWO_LEVEL] }
            self_service_mode:
              {
                type: [string, "null"],
                enum: [SCHOOL_MANAGED, STUDENT_MANAGED, HYBRID, null],
              }
            self_service_min_grade: { type: integer, minimum: 1, maximum: 12 }
            active: { type: boolean }
        - type: object
          additionalProperties: false
          required: [program_id, code, due_at, timezone, workflow_type]
          properties:
            program_id: { type: string, format: uuid }
            code: { type: string, minLength: 2, maxLength: 30 }
            opens_at: { type: [string, "null"], format: date-time }
            due_at: { type: string, format: date-time }
            timezone: { type: string, enum: [Asia/Ho_Chi_Minh] }
            workflow_type: { type: string, enum: [ONE_LEVEL, TWO_LEVEL] }
        - type: object
          additionalProperties: false
          required: [calendar_date, day_type, name]
          properties:
            calendar_date: { type: string, format: date }
            day_type: { type: string, enum: [HOLIDAY, WORKING_DAY] }
            name: { type: string, minLength: 2, maxLength: 150 }
...
    FieldError:
      type: object
      additionalProperties: false
      required: [field, code]
      properties:
        field: { type: string }
        code: { type: string }
        params: { type: object, additionalProperties: true }
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [code, message_key, field_errors, correlation_id]
      properties:
        code: { type: string, example: VERSION_CONFLICT }
        message_key: { type: string, example: errors.versionConflict }
        field_errors:
          { type: array, items: { $ref: "#/components/schemas/FieldError" } }
        correlation_id: { type: string, format: uuid }
...
  responses:
...
    Unauthorized:
      {
        description: Authentication or MFA required,
        content:
          {
            application/json:
              { schema: { $ref: "#/components/schemas/ErrorResponse" } },
          },
      }
    NotFound:
      {
        description: Missing or concealed by scope,
        content:
          {
            application/json:
              { schema: { $ref: "#/components/schemas/ErrorResponse" } },
          },
      }
    Conflict:
      {
        description: "Duplicate, state or version conflict",
        content:
          {
            application/json:
              { schema: { $ref: "#/components/schemas/ErrorResponse" } },
          },
      }
    ValidationFailed:
      {
        description: Business validation failed,
        content:
          {
            application/json:
              { schema: { $ref: "#/components/schemas/ErrorResponse" } },
          },
      }
    PreconditionRequired:
      {
        description: If-Match/version is required,
        content:
          {
            application/json:
              { schema: { $ref: "#/components/schemas/ErrorResponse" } },
          },
      }
...